day1: an image processing640*480
Here I start Neon. An image is added into this project as an example. 
In this example, the image is a two-dimension array by 640*480, and it is filled by random numbers. The purpose is to get a new array by adding each element of this array with its neighbor two elements.
Here are rules:
1. image[i][j] = image[i][j] + image[i][j-1] + image[i][j+1];
2. image[i][j] = image[i][j] + image[i][j+1], for the first element of every row;
3. image[i][j] = image[i][j] + image[i][j-1], for the last element of every row.